src/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ilya Breitburg <ilya.breytburg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
runware = "^0.4.4"
langchain-groq = "^0.2.5"
langchain-anthropic = "^0.3.9"
python-dotenv = "^1.0.1"
langchain-core = "^0.3.45"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
